Se crea objeto basado en clase Sensores y se completan e imprimen sus propiedades
