+++ sensores_esp32/lib/comunicacion/comunicacion.cpp
@@ -1,0 +1,3 @@
+#include "comunicacion.h"
+
+
