+++ sensores_esp32/lib/comunicacion/comunicacion.h
@@ -1,0 +1,12 @@
+#include "Arduino.h"
+
+class Sensores {       // The class
+  public:             // Access specifier
+       
+    String SSID;
+    String Password;
+    String url_broker;
+    String topic;
+    String value;
+
+};
