--- sensores_esp32/lib/toggleLED.cpp
@@ -1,24 +1,0 @@
-#ifndef toggleLED_h
-#define toggleLED_h
-#include "toggleLED.h"
-#endif
-
-
-void toggleLED(void){
-
-    digitalWrite(LED_BUILTIN, HIGH);
-
-
-}
-
-void printScreen(void){
-
-    digitalWrite(LED_BUILTIN, LOW);
-
-}
-
-void pepe(void){
-
-    digitalWrite(LED_BUILTIN, LOW);
-
-}
