--- sensores_esp32/lib/toggleLED.h
@@ -1,6 +1,0 @@
-#include <Arduino.h>
-#define LED_BUILTIN 2
-
-void toggleLED(void);
-void printScreen(void);
-void pepe(void);
