+++ sensores_esp32/lib/toggleLED/toggleLED.cpp
@@ -1,0 +1,24 @@
+#ifndef toggleLED_h
+#define toggleLED_h
+#include "toggleLED.h"
+#endif
+
+
+void toggleLED(void){
+
+    digitalWrite(LED_BUILTIN, HIGH);
+
+
+}
+
+void printScreen(void){
+
+    digitalWrite(LED_BUILTIN, LOW);
+
+}
+
+void pepe(void){
+
+    digitalWrite(LED_BUILTIN, LOW);
+
+}
