+++ sensores_esp32/platformio.ini
@@ -12,3 +12,4 @@
 platform = espressif32
 board = nodemcu-32s
 framework = arduino
+monitor_speed = 115200
