+++ sensores_esp32/src/main.cpp
@@ -1,5 +1,5 @@
 #include <toggleLED.h>
-
+#include <comunicacion.h>
 #include <EEPROM.h>
 
 
@@ -9,8 +9,22 @@
 void setup() {
   // put your main code here, to run repeatedly:
   //delay(10000);
-  //pinMode(LED_BUILTIN, OUTPUT);
+  Serial.begin(115200);
+  pinMode(LED_BUILTIN, OUTPUT);
+  Sensores sensorTemp;  // Create an object
+
+  // Access attributes and set values
+  sensorTemp.value = 25;
+  sensorTemp.SSID = "infiniem";  
+  sensorTemp.Password = "12345678";  
   
+
+  // Print attribute values
+
+  Serial.println(sensorTemp.value);
+  Serial.println(sensorTemp.SSID);
+  Serial.println(sensorTemp.Password);
+    
 }
 
 void loop() {
